// src/pages/manageteam/types.ts
export type DepartmentStats = {
    name: string;
    point: number;
    leader: string;
    dispatch: number;
    assistent: number;
    driver: number;
    score: number;
};

export type Dispatch = {
    name: string;
    age: number
}

export type Assistent = {
    name: string;
}

export type Driver = {
    name: string;
}